Adding Echantillon

+++ src/main/java/com/example/demo/service/Statistique.java
@@ -13,6 +13,6 @@ public interface Statistique {
      * @return le prix des voitures
      * @throws ArithmeticException s'il n'y a pas de voiture
      */
-    public int prix() throws ArithmeticException;
+    public float prix() throws ArithmeticException;
 
 }
